requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.55.0
greenlet==1.1.2
gspread==5.2.0
httplib2==0.20.4
idna==3.3
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.5
PyMySQL==1.0.2
pyparsing==3.0.7
PySheets==1.0.3
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.32
uritemplate==4.1.1
urllib3==1.26.8
